# [RFC] [mini] add fastbump allocator that is more inline-friendly