@Therzok 2019-10-03
@lewurm out of curiosity, what performance characteristics changes does this yield for normal JIT?