@lewurm 2019-10-03
@Therzok The `gc-stress.exe 1000` example from above, with JIT (without AOT, without LLVM):
```
User Time                           : 10805.05 ms
System Time                         : 184.58 ms
Total Time                          : 10989.65 ms
```
Same but with `MONO_FAST_BUMP=1` enabled:
```
System statistics
User Time                           : 9063.14 ms
System Time                         : 178.75 ms
Total Time                          : 9241.90 ms
```

As for the generated code, before:
```
Simple_Test:
0000000000000c30        subq    $0x8, %rsp
0000000000000c34        movq    0x625(%rip), %rax
0000000000000c3b        testl   $0x1, (%rax)
0000000000000c41        je      0xc48
0000000000000c43        callq   _p_1
0000000000000c48        movq    0x6a9(%rip), %rdi
0000000000000c4f        movl    $0x10, %esi
0000000000000c54        callq   _p_3
0000000000000c59        addq    $0x8, %rsp
0000000000000c5d        retq
0000000000000c5e        nop
```
With `MONO_FAST_BUMP=1`:
```
Simple_Test:
0000000000000ba0        subq    $0x28, %rsp
0000000000000ba4        movq    %r12, (%rsp)
0000000000000ba8        movq    %r13, 0x8(%rsp)
0000000000000bad        movq    %r14, 0x10(%rsp)
0000000000000bb2        movq    %r15, 0x18(%rsp)
0000000000000bb7        movq    0x6a2(%rip), %rax
0000000000000bbe        testl   $0x1, (%rax)
0000000000000bc4        je      0xbcb
0000000000000bc6        callq   _p_1  // plt__jit_icall_mono_threads_state_poll
0000000000000bcb        movq    0x726(%rip), %r15
0000000000000bd2        movq    0x6a7(%rip), %rax
0000000000000bd9        callq   *%rax
0000000000000bdb        movq    %rax, %rcx
0000000000000bde        movq    %rax, %r14
0000000000000be1        addq    $0x600, %r14
0000000000000be8        movq    %r14, %rax
0000000000000beb        movq    (%rax), %r13
0000000000000bee        movq    %r13, %rax
0000000000000bf1        movq    %rax, %r12
0000000000000bf4        addq    $0x10, %r12
0000000000000bf8        movq    %r12, %rax
0000000000000bfb        movq    0x608(%rcx), %rcx
0000000000000c02        cmpq    %rcx, %rax
0000000000000c05        jae     0xc13
0000000000000c07        movq    %r12, (%r14)
0000000000000c0a        movq    %r15, (%r13)
0000000000000c0e        movq    %r13, %r15
0000000000000c11        jmp     0xc1e
0000000000000c13        movq    %r15, %rdi
0000000000000c16        callq   _p_3  // plt__jit_icall_ves_icall_object_new_specific
0000000000000c1b        movq    %rax, %r15
0000000000000c1e        movq    %r15, %rax
0000000000000c21        movq    (%rsp), %r12
0000000000000c25        movq    0x8(%rsp), %r13
0000000000000c2a        movq    0x10(%rsp), %r14
0000000000000c2f        movq    0x18(%rsp), %r15
0000000000000c34        addq    $0x28, %rsp
0000000000000c38        retq
0000000000000c39        nopl    (%rax)
```

Please keep in mind that `gc-stress.exe` is really in favor of this change. The real life performance has to be measured yet; currently it crashes when building the BCL, so it needs some more love.