@lewurm 2019-10-25
Closing it, some thoughts:
* The only reason why inlining this fastbump wrapper emits better code than the regular allocation wrapper is that I cheated: It leaves out the set/unset operation for `in_critical_region`. Correctness-wise, I'm not sure about removing `in_critical_region`. Also see https://github.com/mono/mono/issues/17465#issuecomment-544657678
* Not done here, but I also tried to "lift" `tlab_next` and `tlab_temp_end` directly into the TLS, hoping to get rid of one load each. Alas we only have a `TLS_GET` operation in the JIT, so writing back works by having a pointer in the TLS... which is what the status quo looks like. Adding `TLS_SET` to the JIT looks like a bit of more work, so not sure if it's worth.